+++ NEWS.md
@@ -1,5 +1,5 @@
 
-# dev version
+# development version
 
 # callr 3.3.0