+++ NEWS.md
@@ -1,5 +1,5 @@
 
-# development version
+# callr 3.3.1
 
 * `r_session` now avoids creating `data` and `env` objects in the global
   environment of the subprocess.